requirements.txt
aiohttp==3.7.4
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
Automat==20.2.0
certifi==2020.11.8
cffi==1.14.5
chardet==3.0.4
constantly==15.1.0
discord.py==1.7.1
docopt==0.6.2
ffmpeg==1.4
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
multidict==5.1.0
osascript==2020.12.3
pipreqs==0.4.10
psutil==5.8.0
psycopg2==2.8.6
pycparser==2.20
PyNaCl==1.4.0
python-dateutil==2.8.1
requests==2.25.1
runcmd==2020.12.3
schedule==1.0.0
six==1.15.0
Twisted==21.2.0
twitchAPI==2.3.2
typing-extensions==3.7.4.3
urllib3==1.26.3
websockets==9.1
yarg==0.1.9
yarl==1.6.3
youtube-dl==2021.4.26
zope.interface==5.3.0
